Create DummyGraphParser / GraphParser distinction
